+++ libs/react-components/src/tedi/components/buttons/closing-button/closing-button.stories.tsx
@@ -4,13 +4,13 @@ import { Col, Row } from '../../grid';
 import ClosingButton, { ClosingButtonProps } from './closing-button';
 
 /**
- * [Figma ↗](https://www.figma.com/file/jWiRIXhHRxwVdMSimKX2FF/TEDI-Design-System-(draft)?type=designnode-id=4514-63815&m=dev)<br/>
- * [Zeroheight ↗](https://tedi.tehik.ee/1ee8444b7/p/30df1b-closing-button)
+ * <a href="https://www.figma.com/design/jWiRIXhHRxwVdMSimKX2FF/TEDI-READY-(work-in-progress)?node-id=4514-63815&m=dev" target="_BLANK">Figma ↗</a><br/>
+ * <a href="https://tedi.tehik.ee/1ee8444b7/p/30df1b-closing-button" target="_BLANK">Zeroheight ↗</a>
  */
 
 const meta: Meta<typeof ClosingButton> = {
   component: ClosingButton,
-  title: 'Tedi-ready/Components/Buttons/ClosingButton',
+  title: 'Tedi-Ready/Components/Buttons/ClosingButton',
 };
 
 const sizeArray: ClosingButtonProps['size'][] = ['medium', 'large'];