+++ libs/react-components/src/tedi/components/icon/icon.tsx
@@ -1,4 +1,5 @@
 import cn from 'classnames';
+import { forwardRef } from 'react';
 
 import styles from './icon.module.scss';
 
@@ -60,7 +61,7 @@ export interface IconProps {
   background?: IconBackgroundColor;
 }
 
-export const Icon = (props: IconProps): JSX.Element => {
+export const Icon = forwardRef<HTMLDivElement, IconProps>((props: IconProps, ref): JSX.Element => {
   const {
     className,
     name,
@@ -76,6 +77,8 @@ export const Icon = (props: IconProps): JSX.Element => {
   const wrapperBEM = cn(styles['tedi-icon--wrapper'], {
     [styles['tedi-icon--bg']]: background,
     [styles[`tedi-icon--bg-${background}`]]: background,
+    [styles[`tedi-icon--wrapper--size-${size}`]]: size,
+    [styles[`tedi-icon--wrapper--${display}`]]: display,
   });
 
   const iconBEM = cn(
@@ -83,20 +86,28 @@ export const Icon = (props: IconProps): JSX.Element => {
     'material-symbols',
     type && [`material-symbols--${type}`],
     styles['tedi-icon'],
-    display && styles[`tedi-icon--${display}`],
     color && styles[`tedi-icon--color-${color}`],
     size && styles[`tedi-icon--size-${size}`],
+    display && styles[`tedi-icon--${display}`],
     filled && styles['tedi-icon--filled'],
     className
   );
 
-  return (
-    <div className={wrapperBEM}>
-      <span className={iconBEM} data-name="icon" role="img" aria-hidden={true} {...rest}>
-        {name}
-      </span>
-    </div>
+  const iconElement = (
+    <span ref={!background ? ref : null} className={iconBEM} data-name="icon" role="img" aria-hidden={true} {...rest}>
+      {name}
+    </span>
   );
-};
+
+  if (background) {
+    return (
+      <div className={wrapperBEM} ref={ref}>
+        {iconElement}
+      </div>
+    );
+  }
+
+  return iconElement;
+});
 
 Icon.displayName = 'Icon';