+++ libs/react-components/src/tedi/components/navigation/link/link.stories.tsx
@@ -8,8 +8,8 @@ import { VerticalSpacing } from '../../vertical-spacing';
 import { Link, LinkProps } from './link';
 
 /**
- * [Figma ↗](https://www.figma.com/file/jWiRIXhHRxwVdMSimKX2FF/TEDI-Design-System-(draft)?type=design&node-id=2160-25385&m=dev)<br/>
- * [Zeroheight ↗](https://tedi.tehik.ee/1ee8444b7/p/76e0c0-link)
+ * <a href="https://www.figma.com/file/jWiRIXhHRxwVdMSimKX2FF/TEDI-Design-System-(draft)?type=design&node-id=2160-25385&m=dev" target="_BLANK">Figma ↗</a><br/>
+ * <a href="https://tedi.tehik.ee/1ee8444b7/p/76e0c0-link" target="_BLANK">Zeroheight ↗</a>
  */
 
 const meta: Meta<typeof Link> = {
@@ -76,31 +76,27 @@ const TemplateColumn: StoryFn<TemplateMultipleProps> = (args) => {
                 {value}
               </Text>
             </Col>
-            <Col>
-              <div className="display-flex align-items-center gap-3">
-                <Link id={value} href="#" {...buttonProps}>
-                  Create
-                </Link>
-                <Link id={value} href="#" {...buttonProps} iconRight="arrow_right_alt">
-                  Continue
-                </Link>
-                <Link id={value} href="#" {...buttonProps} iconLeft="edit">
-                  Edit
-                </Link>
-              </div>
+            <Col className="display-flex align-items-center gap-3">
+              <Link id={value} href="#" {...buttonProps}>
+                Create
+              </Link>
+              <Link id={value} href="#" {...buttonProps} iconRight="arrow_right_alt">
+                Continue
+              </Link>
+              <Link id={value} href="#" {...buttonProps} iconLeft="edit">
+                Edit
+              </Link>
             </Col>
-            <Col>
-              <div className="display-flex align-items-center gap-3">
-                <Link id={value} href="#" size="small" {...buttonProps}>
-                  Create
-                </Link>
-                <Link id={value} href="#" size="small" {...buttonProps} iconRight="arrow_right_alt">
-                  Continue
-                </Link>
-                <Link id={value} href="#" size="small" {...buttonProps} iconLeft="edit">
-                  Edit
-                </Link>
-              </div>
+            <Col className="display-flex align-items-center gap-3">
+              <Link id={value} href="#" size="small" {...buttonProps}>
+                Create
+              </Link>
+              <Link id={value} href="#" size="small" {...buttonProps} iconRight="arrow_right_alt">
+                Continue
+              </Link>
+              <Link id={value} href="#" size="small" {...buttonProps} iconLeft="edit">
+                Edit
+              </Link>
             </Col>
           </Row>
         ))}
@@ -244,5 +240,13 @@ export const FullWidth = {
     visualType: 'primary',
     children: 'Link that stretches',
     underline: false,
+  },
+};
+
+export const LongText = {
+  args: {
+    href: '#',
+    iconRight: 'north_east',
+    children: 'This is a very long link text that should wrap into multiple lines',
   },
 };