+++ libs/react-components/src/tedi/components/skeleton/skeleton.stories.tsx
@@ -6,8 +6,8 @@ import { VerticalSpacing } from '../../../tedi/components/vertical-spacing';
 import { Skeleton, SkeletonBlock, SkeletonProps } from '.';
 
 /**
- * [Figma ↗](https://www.figma.com/file/jWiRIXhHRxwVdMSimKX2FF/TEDI-Design-System-(draft)?type=design&node-id=2188-34298&m=dev)<br/>
- * [Zeroheight ↗](https://tedi.tehik.ee/1ee8444b7/p/429294-skeleton)
+ * <a href="https://www.figma.com/file/jWiRIXhHRxwVdMSimKX2FF/TEDI-Design-System-(draft)?type=design&node-id=2188-34298&m=dev" target="_BLANK">Figma ↗</a><br/>
+ * <a href="https://tedi.tehik.ee/1ee8444b7/p/429294-skeleton" target="_BLANK">Zeroheight ↗</a>
  */
 
 const meta: Meta<typeof Skeleton> = {